backend/AI_model/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
torch>=2.1.0
numpy>=1.24.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
